requirements.txt
pip
viam-sdk
# boto3
pillow == 10.0.1
requests
httpx
python-dotenv